Enabled combining multiple blocks to one
Because all of them are loaded with the same parameters, not much has to
be done for this. Nev content (Events, SpikeTrains, Units,
(ChannelIndexes for Units)) does not need to be touched. The resulting
Block object will be the Block of the first loaded nsX file, the
Segments have to be the same for all (see BaseFromRawIO). So Epochs and
AnalogSignals are simply appended to the Segments by index (position in
Block.segments). For
ChannelIndexes it's a bit more difficult because the relation does not
get clear by the position in a list but rather ChannelIndex.name.